+++ src/odr/internal/open_strategy.hpp
@@ -5,7 +5,7 @@
 
 namespace odr {
 enum class FileType;
-struct DecodePreference;
+struct DecodeOptions;
 class Logger;
 } // namespace odr
 
@@ -22,12 +22,7 @@ list_file_types(const std::shared_ptr<abstract::File> &file,
                 const Logger &logger);
 
 std::unique_ptr<abstract::DecodedFile>
-open_file(const std::shared_ptr<abstract::File> &file, const Logger &logger);
-std::unique_ptr<abstract::DecodedFile>
-open_file(const std::shared_ptr<abstract::File> &file, FileType as,
-          const Logger &logger);
-std::unique_ptr<abstract::DecodedFile>
 open_file(const std::shared_ptr<abstract::File> &file,
-          const DecodePreference &preference, const Logger &logger);
+          const DecodeOptions &options, const Logger &logger);
 
 } // namespace odr::internal::open_strategy