+++ test/src/internal/oldms/xls_test.cpp
@@ -300,7 +300,7 @@ TEST(OldMs, xls_empty) {
   const Logger logger = Logger::create_stdio("odr-test", LogLevel::verbose);
 
   const DocumentFile document_file =
-      open(TestData::test_file_path("odr-public/xls/empty.xls"), logger)
+      open(TestData::test_file_path("odr-public/xls/empty.xls"), {}, logger)
           .as_document_file();
 
   EXPECT_EQ(document_file.file_type(), FileType::legacy_excel_worksheets);
@@ -326,7 +326,7 @@ TEST(OldMs, xls_file_example_10) {
 
   const DocumentFile document_file =
       open(TestData::test_file_path("odr-public/xls/file_example_XLS_10.xls"),
-           logger)
+           {}, logger)
           .as_document_file();
 
   EXPECT_EQ(document_file.file_type(), FileType::legacy_excel_worksheets);
@@ -371,7 +371,7 @@ TEST(OldMs, xls_file_example_5000) {
 
   const DocumentFile document_file =
       open(TestData::test_file_path("odr-public/xls/file_example_XLS_5000.xls"),
-           logger)
+           {}, logger)
           .as_document_file();
 
   const Document document = document_file.document();