+++ test/src/internal/ooxml/ooxml_spreadsheet_merge_test.cpp
@@ -73,9 +73,10 @@ Sheet first_sheet(const Document &document) {
 }
 
 Document decode(const std::shared_ptr<abstract::File> &file) {
-  return Document(DecodedFile(open_strategy::open_file(file, Logger::null()))
-                      .as_document_file()
-                      .document());
+  return Document(
+      DecodedFile(open_strategy::open_file(file, {}, Logger::null()))
+          .as_document_file()
+          .document());
 }
 
 constexpr const char *two_cells =